requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
opencv-python==4.8.0.76

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning Utilities
scikit-learn==1.3.0
tqdm==4.66.1

# Web Application
streamlit==1.27.0

# Logging
tensorboard==2.14.0

# Dataset Download (Optional)
kaggle==1.5.16

# Image Augmentation (Optional but recommended)
albumentations==1.3.1
